Extracted FilterUnit and PopoverFilter components
