+++ src/lib/list/ui/index.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
-import { Filter } from '../../types/filter';
 import Alert from '../../components/alert';
 import { paginationBoundaries } from '../order-utils';
-import { SearchUnit } from './form';
 import GlobalSearch from './global-search';
+import PopoverFilter from './popover-filter';
+import FilterUnit from './filter-unit';
 
 import {
   Table,
@@ -12,17 +12,14 @@ import {
   TableCell,
   TableRow,
   TableHead,
-  TableContainer,
-  Popover,
-  IconButton
+  TableContainer
 } from '@material-ui/core';
 
 import {
   KeyboardArrowUp as ChevronUp,
   KeyboardArrowDown as ChevronDown,
   ArrowUpward,
-  ArrowDownward,
-  FilterList as FilterListIcon
+  ArrowDownward
 } from '@material-ui/icons';
 
 interface NoRowProps {
@@ -97,169 +94,6 @@ export const HeaderUnit = (props: HeaderUnitProps): JSX.Element => {
   const { children } = props;
 
   return <TableCell style={{ fontWeight: 'bold' }}>{children}</TableCell>;
-};
-
-interface PopoverFilterProps {
-  children: React.ReactNode | JSX.Element;
-}
-
-export const PopoverFilter = (props: PopoverFilterProps): JSX.Element => {
-  const [anchorEl, setAnchorEl] = React.useState<HTMLButtonElement | null>(
-    null
-  );
-
-  const handleClick = (event: React.MouseEvent<HTMLButtonElement>): void => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = (): void => {
-    setAnchorEl(null);
-  };
-
-  const open = Boolean(anchorEl);
-
-  const { children } = props;
-
-  return (
-    <>
-      <IconButton onClick={handleClick}>
-        <FilterListIcon />
-      </IconButton>
-      <Popover
-        open={open}
-        anchorEl={anchorEl}
-        onClose={handleClose}
-        anchorOrigin={{
-          vertical: 'bottom',
-          horizontal: 'center'
-        }}
-        transformOrigin={{
-          vertical: 'top',
-          horizontal: 'center'
-        }}
-      >
-        {children}
-      </Popover>
-    </>
-  );
-};
-
-interface CheckboxInputValue<A> {
-  name: keyof A;
-  value: {
-    func: Function;
-    value: number | string;
-  };
-}
-
-interface FilterUnitProps<A> {
-  filter?: boolean | Filter<A>;
-  filters: any;
-  name: keyof A;
-  onChange: (
-    inputValue:
-      | { name: keyof A; value: any; type?: string }
-      | CheckboxInputValue<A>
-  ) => void;
-}
-
-export const FilterUnit = <A,>(
-  props: FilterUnitProps<A>
-): JSX.Element | null => {
-  const { filter, filters, name, onChange } = props;
-
-  if (typeof filter === 'boolean' && filter === true) {
-    return (
-      <PopoverFilter>
-        <SearchUnit
-          name={name}
-          value={filters[name]}
-          onChange={(v): void => onChange({ name, value: v.value })}
-        />
-      </PopoverFilter>
-    );
-  }
-
-  if (typeof filter === 'object' && filter.type === 'string') {
-    return (
-      <PopoverFilter>
-        <SearchUnit
-          name={name}
-          value={filters[name] ? filters[name].value.value : ''}
-          onChange={(v): void => {
-            onChange({
-              name,
-              value: { value: v.value, func: filter.func }
-            });
-          }}
-        />
-      </PopoverFilter>
-    );
-  }
-
-  if (typeof filter === 'object' && Array.isArray(filter.options)) {
-    if (filter.type === 'category') {
-      const v = filters[name] ? filters[name].value : [];
-
-      return (
-        <PopoverFilter>
-          {filter.options.map((option, i) => (
-            <span key={i}>
-              <input
-                checked={v.includes(option.key)}
-                type="checkbox"
-                onChange={(): void =>
-                  onChange({
-                    name,
-                    value: {
-                      value: option.key,
-                      func: filter.func
-                    },
-                    type: filter.type
-                  })
-                }
-              />{' '}
-              {option.value}
-              <br />
-            </span>
-          ))}
-        </PopoverFilter>
-      );
-    }
-
-    if (filter.type === 'select') {
-      return (
-        <PopoverFilter>
-          {filter.options.map((option, i) => (
-            <span key={i}>
-              <input
-                checked={
-                  filters[name]
-                    ? filters[name].value.value === option.key
-                    : false
-                }
-                type="radio"
-                onChange={(): void =>
-                  onChange({
-                    name,
-                    value: {
-                      value: option.key,
-                      func: filter.func
-                    },
-                    type: filter.type
-                  })
-                }
-              />{' '}
-              {option.value}
-              <br />
-            </span>
-          ))}
-        </PopoverFilter>
-      );
-    }
-  }
-
-  return null;
 };
 
 interface OrderControllerUpAndDownProps {
@@ -370,4 +204,4 @@ export const RecordInfo = (props: RecordInfoProps): JSX.Element | null => {
   );
 };
 
-export { GlobalSearch };
+export { FilterUnit, GlobalSearch, PopoverFilter };
